requirements-no-torch.txt
einops
h5py
numpy
pandas
matplotlib
seaborn
Pillow
h5py
scikit_learn
scipy
tqdm
beautifulsoup4
nltk
datasets
tokenizers==0.12.1
easydict
requests
bib-lookup>=0.0.19
deprecate-kwargs
deprecated
torch-ecg>=0.0.26
pytest
